- working with dart 2.6.1
- update dependencies
- re-apply analysis options strong mode

- bezier_curve2.dart update position of the points to move them from outside

+++ webgl/webgl/lib/src/webgl_objects/webgl_rendering_context.dart
@@ -575,7 +575,7 @@ class WebGLRenderingContext{
         bitmask |= mask as int;
       }
     }else {
-      bitmask = masks;
+      bitmask = masks as int;
     }
     
     gl.clear(bitmask);
